// react-module/src/openapi/models/atomic-product-template-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicProductTemplateDto
 */
export interface AtomicProductTemplateDto {
    /**
     * 
     * @type {number}
     * @memberof AtomicProductTemplateDto
     */
    id?: any;
    /**
     * 
     * @type {Date}
     * @memberof AtomicProductTemplateDto
     */
    createdAt?: any;
    /**
     * 
     * @type {Date}
     * @memberof AtomicProductTemplateDto
     */
    updatedAt?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductTemplateDto
     */
    createdBy?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductTemplateDto
     */
    modifiedBy?: any;
    /**
     * 
     * @type {boolean}
     * @memberof AtomicProductTemplateDto
     */
    deleted?: any;
    /**
     * 
     * @type {AtomicProductDefinition}
     * @memberof AtomicProductTemplateDto
     */
    atomicProductDefinition?: any;
    /**
     * 
     * @type {LayerShape}
     * @memberof AtomicProductTemplateDto
     */
    layerShape?: any;
    /**
     * 
     * @type {number}
     * @memberof AtomicProductTemplateDto
     */
    layerNumber?: any;
    /**
     * 
     * @type {Array&lt;AtomicProduct&gt;}
     * @memberof AtomicProductTemplateDto
     */
    internalProducts?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductTemplateDto
     */
    title?: any;
}